@@ -12,18 +12,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.8",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.8",
-    "@angular/compiler": "^5.2.8",
-    "@angular/core": "^5.2.8",
-    "@angular/forms": "^5.2.8",
-    "@angular/http": "^5.2.8",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.8",
-    "@angular/platform-browser-dynamic": "^5.2.8",
-    "@angular/platform-server": "^5.2.8",
-    "@angular/router": "^5.2.8",
+    "@angular/animations": "6.0.5",
+    "@angular/cdk": "^6.0.0",
+    "@angular/common": "6.0.5",
+    "@angular/compiler": "6.0.5",
+    "@angular/core": "6.0.5",
+    "@angular/forms": "6.0.5",
+    "@angular/http": "6.0.5",
+    "@angular/material": "^6.0.0",
+    "@angular/platform-browser": "6.0.5",
+    "@angular/platform-browser-dynamic": "6.0.5",
+    "@angular/platform-server": "6.0.5",
+    "@angular/router": "6.0.5",
     "@types/lodash": "^4.14.87",
     "accounting": "^0.4.1",
     "core-js": "^2.4.1",
@@ -32,17 +32,17 @@
     "html2canvas": "^0.5.0-beta4",
     "jszip": "^3.1.4",
     "point-in-polygon": "^1.0.1",
-    "rxjs": "^5.5.6",
+    "rxjs": "^6.2.1",
     "syd-tessellation": "0.0.66",
     "syd-tiling": "0.0.15",
     "syd-visualization": "0.0.85",
-    "zone.js": "^0.8.4",
-    "xml-writer": "1.7.0"
+    "xml-writer": "1.7.0",
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.3",
-    "@angular/compiler-cli": "^5.2.8",
-    "@angular/language-service": "4.4.3",
+    "@angular/cli": "6.0.8",
+    "@angular/compiler-cli": "6.0.5",
+    "@angular/language-service": "6.0.5",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.60",
@@ -60,6 +60,7 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": ">=2.4.2 <2.7.0"
+    "typescript": "2.7.2",
+    "@angular-devkit/build-angular": "~0.6.8"
   }
 }
